Added Vue Router as well as removed memory leak of socket io events
The actions addMessage and removeMessage create event functions that
never get removed.  In other words if a route changes or a hot reload
occurs (component reloads) then new addMessage listeners and
removeMessage listeners will be created; and the previous ones are not
removed. This leads to a memory leak as well as duplicate messages being
stored in the state.  This causes issues with rendering as well for the
v-for list.